requirements.txt
crawl4ai
pydantic
python-dotenv
playwright 